Fix missing storm nodes in print mode
