+++ frontend/src/components/MapView/Layers/AnticipatoryActionStormLayer/index.tsx
@@ -23,9 +23,9 @@ import { updateDateRange } from 'context/mapStateSlice';
 import { useWindStatesByTime } from 'components/MapView/DateSelector/TimelineItems/hooks';
 import { getAAColor } from 'components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionStormPanel/utils';
 import { AACategory } from 'context/anticipatoryAction/AAStormStateSlice/parsedStormDataTypes';
-import anticipatoryActionIcons from 'components/Common/AnticipatoryAction/icons';
 import { AAStormTimeSeriesFeature, TimeSeries } from 'prism-common/';
 import maplibregl from 'maplibre-gl';
+import { WIND_TYPE_TO_ICON_MAP, loadStormIcons } from './constants';
 import AAStormDatePopup from './AAStormDatePopup';
 import AAStormLandfallPopup from './AAStormLandfallPopup';
 
@@ -39,25 +39,6 @@ interface AnticipatoryActionStormLayerProps {
 
 // Use admin level 2 boundary layer
 const boundaryLayer = getBoundaryLayersByAdminLevel(2);
-
-const WIND_TYPE_TO_ICON_MAP: Record<string, string> = {
-  disturbance: anticipatoryActionIcons.disturbance,
-  'tropical-disturbance': anticipatoryActionIcons.disturbance,
-  low: anticipatoryActionIcons.disturbance,
-  'tropical-depression': anticipatoryActionIcons.tropicalDepression,
-  'post-tropical-depression': anticipatoryActionIcons.postTropicalDepression,
-  'sub-tropical-depression': anticipatoryActionIcons.subTropicalDepression,
-  'extratropical-system': anticipatoryActionIcons.extraTropicalSystem,
-  'moderate-tropical-storm': anticipatoryActionIcons.moderateStorm,
-  'severe-tropical-storm': anticipatoryActionIcons.severeTropicalStorm,
-  'tropical-cyclone': anticipatoryActionIcons.tropicalCyclone,
-  'intense-tropical-cyclone': anticipatoryActionIcons.intenseTropicalCyclone,
-  'very-intense-tropical-cyclone':
-    anticipatoryActionIcons.veryIntenseTropicalCyclone,
-  inland: anticipatoryActionIcons.inland,
-  dissipating: anticipatoryActionIcons.dissipating,
-  default: anticipatoryActionIcons.default,
-};
 
 const AnticipatoryActionStormLayer = React.memo(
   ({ layer }: AnticipatoryActionStormLayerProps) => {
@@ -261,20 +242,7 @@ const AnticipatoryActionStormLayer = React.memo(
 
     // Load all images from the mapping
     const loadImages = useCallback(() => {
-      const loadImage = (url: string, name: string) => {
-        map?.loadImage(url, (error, image) => {
-          if (error) {
-            throw error;
-          }
-          if (!map.hasImage(name)) {
-            map.addImage(name, image!);
-          }
-        });
-      };
-
-      Object.entries(WIND_TYPE_TO_ICON_MAP).forEach(([name, url]) => {
-        loadImage(url, name);
-      });
+      loadStormIcons(map, true); // Throw on error for main map
     }, [map]);
 
     useEffect(() => {
