+++ frontend/src/components/NavBar/PrintImage/printPreview.tsx
@@ -14,6 +14,7 @@ import { Panel, AdminLevelDataLayerProps } from 'config/types';
 import useLayers from 'utils/layers-utils';
 import { addFillPatternImagesInMap } from 'components/MapView/Layers/AdminLevelDataLayer/utils';
 import { mapStyle } from 'components/MapView/Map/utils';
+import { loadStormIcons } from 'components/MapView/Layers/AnticipatoryActionStormLayer/constants';
 import {
   dateRangeSelector,
   mapSelector,
@@ -273,6 +274,9 @@ function PrintPreview() {
                         ),
                       ),
                     );
+
+                    // Load storm icons for anticipatory action storm layers
+                    loadStormIcons(mapRef.current?.getMap(), false); // Don't throw on error for print preview
                   }}
                   mapStyle={selectedMapStyle || mapStyle.toString()}
                   maxBounds={selectedMap.getMaxBounds() ?? undefined}
